Move container components, link todo list
