+++ src/components/App.js
@@ -1,12 +1,14 @@
 import React, { Component } from 'react';
-import AddTodoForm from './AddTodoForm';
+import ToDoList from '../containers/ToDoList';
+import AddTodoForm from '../containers/AddTodoForm';
 
 class App extends Component {
   render() {
     return (
       <div className="App">
-        <h1>Another Todo Box</h1>
+        <h1>Todo Box</h1>
         <AddTodoForm />
+        <ToDoList />
       </div>
     );
   }
