--- src/components/ToDo.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const ToDo = ({ id, text, completed }) => {
-  return <li>{text}</li>;
-};
-
-export default ToDo;
