+++ src/containers/ToDo.js
@@ -1,0 +1,8 @@
+import React from 'react';
+// import { connect } from 'react-redux';
+
+const ToDo = ({ todo }) => {
+  return <li>{todo}</li>;
+};
+
+export default ToDo;
